Fix: make raspi3.yaml work with current vmdb2
Add the kpartx step. It is now mandtory for building an image
file. (Not necessary if building on a real drive directly.)

Use only one set of tag names: boot-part and root-fs. vmdb2 has
changed so that partitions and mount points share the same tag name
space: there's no need to have a tag for the partition and for the
mount point for that partition, anymore. As part of that, there's no
need to specify both the partition tag for the root partition, and
then also the tag for the mount point for that partition.

ea7b9bc1d3e5de362b5edf12603ee83c708ad1a1 is the commit of vmdb2 that
this works with.