Add: support for caching and unpacking the rootfs
This speeds up iterative development of a vmdb specification file: the
output of qemu-debootstrap and the apt installation additinal packages
are cached. They take up most of the time of a full run. If modifying
other parts of raspi3.yaml, this means an iteration drops from about
20 minutes to 16 seconds on my laptop.

If making changes to the cached parts, the cache tarball needs to be
deleted between runs.

The --rootfs-tarball options MUST now be used on all invocations.